+++ src/components/AlertSnackbar/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { Snackbar } from '@material-ui/core';
+import { Snackbar, Typography, makeStyles } from '@material-ui/core';
 
 import MuiAlert from '@material-ui/lab/Alert';
 import { hideAlertSnackbar } from 'ducks';
@@ -10,7 +10,15 @@ const Alert = (props) => {
   return <MuiAlert elevation={6} variant="filled" {...props} />;
 };
 
+const useStyles = makeStyles((theme) => ({
+  message: {
+    fontWeight: 'bold',
+    color: theme.palette.white,
+  },
+}));
+
 const AlertSnackbar = () => {
+  const classes = useStyles();
   const dispatch = useDispatch();
   const { severity, message, open, options } = useSelector(
     (state) => state.alertSnackbar
@@ -31,7 +39,9 @@ const AlertSnackbar = () => {
       onClose={handleCloseAlertSnackbar}
     >
       <Alert onClose={handleCloseAlertSnackbar} severity={severity}>
-        {message}
+        <Typography variant="subtitle2" className={classes.message}>
+          {message}
+        </Typography>
       </Alert>
     </Snackbar>
   );