+++ src/components/Spinner/index.js
@@ -8,6 +8,7 @@ const Spinner = () => {
         display: 'flex',
         justifyContent: 'center',
         alignItems: 'center',
+        height: '100%',
       }}
     >
       <CircularProgress