+++ src/ducks/alertSnackbar.js
@@ -1,11 +1,8 @@
-const SHOW_ALERT_SNACKBAR = 'rhoades/alertSnackbar/SHOW_ALERT_SNACKBAR';
-const HIDE_ALERT_SNACKBAR = 'rhoades/alertSnackbar/HIDE_ALERT_SNACKBAR';
-
-const initialState = {
-  open: false,
-  severity: '',
-  message: '',
-  options: {},
+// const SHOW_ALERT_SNACKBAR = 'rhoades/alertSnackbar/SHOW_ALERT_SNACKBAR';
+// const HIDE_ALERT_SNACKBAR = 'rhoades/alertSnackbar/HIDE_ALERT_SNACKBAR';
+export const ALERT_SNACKBAR = {
+  SHOW: 'rhoades/alertSnackbar/SHOW_ALERT_SNACKBAR',
+  HIDE: 'rhoades/alertSnackbar/HIDE_ALERT_SNACKBAR',
 };
 
 // default options of the alert snackbar
@@ -14,9 +11,16 @@ const defaultOptions = {
   duration: 1200,
 };
 
+const initialState = {
+  open: false,
+  severity: '',
+  message: '',
+  options: defaultOptions,
+};
+
 export default function reducer(state = initialState, action) {
   switch (action.type) {
-    case SHOW_ALERT_SNACKBAR:
+    case ALERT_SNACKBAR.SHOW:
       return {
         ...state,
         open: true,
@@ -24,7 +28,7 @@ export default function reducer(state = initialState, action) {
         message: action.payload.message,
         options: action.payload.options,
       };
-    case HIDE_ALERT_SNACKBAR:
+    case ALERT_SNACKBAR.HIDE:
       return {
         ...state,
         open: false,
@@ -35,14 +39,14 @@ export default function reducer(state = initialState, action) {
 }
 
 export const showAlertSnackbar = (
-  severity = '',
-  message = '',
+  severity,
+  message,
   options = defaultOptions
 ) => ({
-  type: SHOW_ALERT_SNACKBAR,
+  type: ALERT_SNACKBAR.SHOW,
   payload: { severity, message, options },
 });
 
 export const hideAlertSnackbar = () => ({
-  type: HIDE_ALERT_SNACKBAR,
+  type: ALERT_SNACKBAR.HIDE,
 });