+++ src/tools/request.js
@@ -1,0 +1,34 @@
+import axios from 'axios';
+const baseURL = process.env.REACT_APP_API_URL;
+
+const axiosInstance = (headers) => {
+  let instance = axios.create({
+    baseURL,
+    mode: 'no-cors',
+    headers,
+  });
+
+  return instance;
+};
+
+const Get = async (route, params = {}, headers = {}) => {
+  try {
+    const { data } = await axiosInstance(headers).get(route, {
+      params,
+    });
+    return data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const Post = async (route, json = {}, headers = {}) => {
+  try {
+    const { data } = await axiosInstance(headers).post(route, json);
+    return data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+export { Get, Post };